fixed build warning

+++ kernel/fs/mminode.c
@@ -140,7 +140,7 @@ void fs_inode_sync_physical_page(struct inode *node, addr_t virt, size_t offset)
 	size_t len = PAGE_SIZE;
 	if(len + offset > (size_t)node->len)
 		len = node->len - offset;
-	if(offset >= node->len)
+	if(offset >= (size_t)node->len)
 		return;
 	if(node->i_ops && vfs_write_inode(node, offset, len, (void *)virt) < 0)
 		printk(0, "[mminode]: warning: failed to write back data\n");
